Configure jest/enzyme and initialize files for tests
