+++ .gitignore
@@ -5,7 +5,6 @@
 npm
 npm-debug.log
 node_modules/
-public/dist/bundle.js
 pids
 *.pid
 logs
@@ -22,4 +21,9 @@ coverage
 
 # Compiled Dirs (http://nodejs.org/api/addons.html)
 build/
+public/dist/bundle.js
+v8-compile-cache-0/
+.babelrc
+.eslintrc.js
+.nvmrc
