--- client/src/components/WelcomeMessage.jsx
@@ -1,28 +1,0 @@
-
-import React, { Fragment } from 'react';
-
-const styles = {
-  heading: {
-    color: '#fff',
-    textAlign: 'center',
-    marginBottom: theme.margin * 2,
-  },
-  logo: {
-    width: 250,
-    heading: 250,
-    objectFit: 'cover',
-  },
-};
-
-const WelcomeMessage = (  ) => {
-  return (
-    <Fragment>
-      <h1 style={}>
-        Welcome To
-      </h1>
-      <img src='' alt="app logo"/>
-    </Fragment>
-  );
-};
-
-export default WelcomeMessage;
