+++ client/src/components/jest.config.js
@@ -1,6 +1,7 @@
 
 // For a detailed explanation regarding each configuration property, visit:
-// https://jestjs.io/docs/en/configuration.html
+// https://jestjs.io/docs/en/configuration.html\
+
 
 module.exports = {
   // Automatically clear mock calls and instances between every test
